my first commit

+++ frontend/src/screens/OrdersScreen.js
@@ -21,10 +21,12 @@ function OrdersScreen(props) {
      
     };
   }, [dispatch,successDelete]);
-
   const deleteHandler = (order) => {
-    dispatch(deleteOrder(order._id));
+    if (window.confirm('Are you sure to delete?')) {
+      dispatch(deleteOrder(order._id));
+    }
   }
+ 
   return loading ? <div>Loading...</div> :
     <div className="content content-margined">
 
@@ -62,8 +64,15 @@ function OrdersScreen(props) {
                 </td>
               <td>{order.deliveredAt}</td>
               <td>
-                <Link to={"/order/" + order._id} className="button secondary" >Details</Link>
-                {' '}
+              <button
+                    type="button"
+                    className="small"
+                    onClick={() => {
+                      props.history.push(`/order/${order._id}`);
+                    }}
+                  >
+                    Details
+                  </button>
                 <button type="button" onClick={() => deleteHandler(order)} className="button secondary">Delete</button>
               </td>
             </tr>))}
